Remove serviceSetupV2 feature flag.
