+++ feature-flags.json
@@ -2,7 +2,6 @@
 	"adsenseSetupV2",
 	"dashboardSharing",
 	"ideaHubModule",
-	"serviceSetupV2",
 	"swgModule",
 	"userInput",
 	"unifiedDashboard",
