+++ includes/Core/Authentication/Setup.php
@@ -14,6 +14,7 @@ use Google\Site_Kit\Context;
 use Google\Site_Kit\Core\Authentication\Clients\OAuth_Client;
 use Google\Site_Kit\Core\Authentication\Exception\Exchange_Site_Code_Exception;
 use Google\Site_Kit\Core\Authentication\Exception\Missing_Verification_Exception;
+use Google\Site_Kit\Core\Permissions\Permissions;
 use Google\Site_Kit\Core\Storage\User_Options;
 use Google\Site_Kit\Core\Util\Feature_Flags;
 
@@ -24,7 +25,7 @@ use Google\Site_Kit\Core\Util\Feature_Flags;
  * @access private
  * @ignore
  */
-abstract class Setup {
+class Setup {
 
 	/**
 	 * Context instance.
@@ -97,7 +98,145 @@ abstract class Setup {
 	 *
 	 * @since 1.48.0
 	 */
-	abstract public function register();
+	public function register() {
+		add_action( 'admin_action_' . Google_Proxy::ACTION_SETUP_START, array( $this, 'handle_action_setup_start' ) );
+		add_action( 'admin_action_' . Google_Proxy::ACTION_VERIFY, array( $this, 'handle_action_verify' ) );
+		add_action( 'admin_action_' . Google_Proxy::ACTION_EXCHANGE_SITE_CODE, array( $this, 'handle_action_exchange_site_code' ) );
+	}
+
+	/**
+	 * Handles the setup start action, taking the user to the proxy setup screen.
+	 *
+	 * @since 1.48.0
+	 */
+	public function handle_action_setup_start() {
+		$nonce        = $this->context->input()->filter( INPUT_GET, 'nonce', FILTER_SANITIZE_STRING );
+		$redirect_url = $this->context->input()->filter( INPUT_GET, 'redirect', FILTER_SANITIZE_URL );
+
+		$this->verify_nonce( $nonce, Google_Proxy::ACTION_SETUP_START );
+
+		if ( ! current_user_can( Permissions::SETUP ) ) {
+			wp_die( esc_html__( 'You have insufficient permissions to connect Site Kit.', 'google-site-kit' ) );
+		}
+
+		if ( ! $this->credentials->using_proxy() ) {
+			wp_die( esc_html__( 'Site Kit is not configured to use the authentication proxy.', 'google-site-kit' ) );
+		}
+
+		$required_scopes = $this->authentication->get_oauth_client()->get_required_scopes();
+		$this->google_proxy->with_scopes( $required_scopes );
+
+		$oauth_setup_redirect = $this->credentials->has()
+			? $this->google_proxy->sync_site_fields( $this->credentials, 'sync' )
+			: $this->google_proxy->register_site( 'sync' );
+
+		if ( is_wp_error( $oauth_setup_redirect ) || ! filter_var( $oauth_setup_redirect, FILTER_VALIDATE_URL ) ) {
+			wp_die( esc_html__( 'The request to the authentication proxy has failed. Please, try again later.', 'google-site-kit' ) );
+		}
+
+		if ( $redirect_url ) {
+			$this->user_options->set( OAuth_Client::OPTION_REDIRECT_URL, $redirect_url );
+		}
+
+		wp_safe_redirect( $oauth_setup_redirect );
+		exit;
+	}
+
+	/**
+	 * Handles the action for verifying site ownership.
+	 *
+	 * @since 1.48.0
+	 * @since 1.49.0 Sets the `verify` and `verification_method` and `site_id` query params.
+	 */
+	public function handle_action_verify() {
+		$input               = $this->context->input();
+		$step                = $input->filter( INPUT_GET, 'step', FILTER_SANITIZE_STRING );
+		$nonce               = $input->filter( INPUT_GET, 'nonce', FILTER_SANITIZE_STRING );
+		$code                = $input->filter( INPUT_GET, 'googlesitekit_code', FILTER_SANITIZE_STRING );
+		$site_code           = $input->filter( INPUT_GET, 'googlesitekit_site_code', FILTER_SANITIZE_STRING );
+		$verification_token  = $input->filter( INPUT_GET, 'googlesitekit_verification_token', FILTER_SANITIZE_STRING );
+		$verification_method = $input->filter( INPUT_GET, 'googlesitekit_verification_token_type', FILTER_SANITIZE_STRING );
+
+		$this->verify_nonce( $nonce );
+
+		if ( ! current_user_can( Permissions::SETUP ) ) {
+			wp_die( esc_html__( 'You don\'t have permissions to set up Site Kit.', 'google-site-kit' ), 403 );
+		}
+
+		if ( ! $code ) {
+			wp_die( esc_html__( 'Invalid request.', 'google-site-kit' ), 400 );
+		}
+
+		if ( ! $verification_token || ! $verification_method ) {
+			wp_die( esc_html__( 'Verifying site ownership requires a token and verification method.', 'google-site-kit' ), 400 );
+		}
+
+		$this->handle_verification( $verification_token, $verification_method );
+
+		$proxy_query_params = array(
+			'step'                => $step,
+			'verify'              => 'true',
+			'verification_method' => $verification_method,
+		);
+
+		// If the site does not have a site ID yet, a site code will be passed.
+		// Handling the site code here will save the extra redirect from the proxy if successful.
+		if ( $site_code ) {
+			try {
+				$this->handle_site_code( $code, $site_code );
+			} catch ( Missing_Verification_Exception $exception ) {
+				$proxy_query_params['site_code'] = $site_code;
+
+				$this->redirect_to_proxy( $code, $proxy_query_params );
+			} catch ( Exchange_Site_Code_Exception $exception ) {
+				$this->redirect_to_splash();
+			}
+		}
+
+		$credentials                   = $this->credentials->get();
+		$proxy_query_params['site_id'] = ! empty( $credentials['oauth2_client_id'] ) ? $credentials['oauth2_client_id'] : '';
+
+		$this->redirect_to_proxy( $code, $proxy_query_params );
+	}
+
+	/**
+	 * Handles the action for exchanging the site code for site credentials.
+	 *
+	 * This action will only be called if the site code failed to be handled
+	 * during the verification step.
+	 *
+	 * @since 1.48.0
+	 */
+	public function handle_action_exchange_site_code() {
+		$input     = $this->context->input();
+		$step      = $input->filter( INPUT_GET, 'step', FILTER_SANITIZE_STRING );
+		$nonce     = $input->filter( INPUT_GET, 'nonce', FILTER_SANITIZE_STRING );
+		$code      = $input->filter( INPUT_GET, 'googlesitekit_code', FILTER_SANITIZE_STRING );
+		$site_code = $input->filter( INPUT_GET, 'googlesitekit_site_code', FILTER_SANITIZE_STRING );
+
+		$this->verify_nonce( $nonce );
+
+		if ( ! current_user_can( Permissions::SETUP ) ) {
+			wp_die( esc_html__( 'You don\'t have permissions to set up Site Kit.', 'google-site-kit' ), 403 );
+		}
+
+		if ( ! $code || ! $site_code ) {
+			wp_die( esc_html__( 'Invalid request.', 'google-site-kit' ), 400 );
+		}
+
+		try {
+			$this->handle_site_code( $code, $site_code );
+		} catch ( Missing_Verification_Exception $exception ) {
+			$this->redirect_to_proxy( $code, compact( 'site_code', 'step' ) );
+		} catch ( Exchange_Site_Code_Exception $exception ) {
+			$this->redirect_to_splash();
+		}
+
+		$credentials = $this->credentials->get();
+		$site_id     = ! empty( $credentials['oauth2_client_id'] ) ? $credentials['oauth2_client_id'] : '';
+
+		$this->redirect_to_proxy( $code, compact( 'site_id', 'step' ) );
+	}
 
 	/**
 	 * Verifies the given nonce for a setup action.
@@ -184,13 +323,8 @@ abstract class Setup {
 	 * @param array  $params Additional query parameters to include in the proxy redirect URL.
 	 */
 	protected function redirect_to_proxy( $code = '', $params = array() ) {
-		if ( Feature_Flags::enabled( 'serviceSetupV2' ) ) {
-			$params['code'] = $code;
-			$url            = $this->authentication->get_google_proxy()->setup_url_v2( $params );
-		} else {
-			$url = $this->authentication->get_oauth_client()->get_proxy_setup_url( $code );
-			$url = add_query_arg( $params, $url );
-		}
+		$params['code'] = $code;
+		$url            = $this->authentication->get_google_proxy()->setup_url_v2( $params );
 
 		wp_safe_redirect( $url );
 		exit;
